Add typing for classes.
Scrap the last commit message, the reason the error was missed was due to the any type being used here.

Swapped to use a check variable now. Ideally, we could use Optional and define the variables as None in init, but I'm not sure how to let mypy know about the check for "is not None".